% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesigmoid.R
\name{aesigmoid}
\alias{aesigmoid}
\title{Asymmetric sigmoid helper}
\usage{
aesigmoid(t, onset, duration)
}
\arguments{
\item{t}{Time value.}

\item{onset}{Onset in days.}

\item{duration}{Duration in days.}
}
\value{
Probability value between 0 and 1.
}
\description{
Computes a sigmoidal curve used for within herd state transitions.
}
