% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_files.R
\name{aggregate_files}
\alias{aggregate_files}
\title{Aggregate simulation files}
\usage{
aggregate_files(data_path, sim_type = "R2R", sample_percentage = 100)
}
\arguments{
\item{data_path}{Directory containing result files.}

\item{sim_type}{Either `"R2R"` or `"herd"` to choose file patterns.}

\item{sample_percentage}{Percentage of files to randomly sample.}
}
\value{
A combined `data.frame` of all sampled records.
}
\description{
Read a selection of simulation output CSV files and combine them into a
single data frame.
}
