% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_no_infection_percent.R
\name{calculate_no_infection_percent}
\alias{calculate_no_infection_percent}
\title{Proportion of trials with no infections by room}
\usage{
calculate_no_infection_percent(data_path)
}
\arguments{
\item{data_path}{Directory of result CSV files.}
}
\value{
A data frame of rooms and the percentage with no infections.
}
\description{
For room-to-room simulations compute the percentage of trials in which no
infections occurred in specified rooms.
}
