% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_phase_stats.R
\name{calculate_phase_stats}
\alias{calculate_phase_stats}
\title{Summarise infection phase durations}
\usage{
calculate_phase_stats(melted_data, units = "hours")
}
\arguments{
\item{melted_data}{Output of `calculate_phase_durations()`.}

\item{units}{Unit of time to report, "hours" or "days".}
}
\value{
A data frame of summary statistics.
}
\description{
Takes the long format data from `calculate_phase_durations()` and produces
summary statistics for each group and phase.
}
