% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{challengeABM-package}
\alias{challengeABM}
\alias{challengeABM-package}
\title{challengeABM: ABM for Simulation of Within-host Virus Dynamics}
\description{
Simulation of virus nasal shedding and blood serum replication witin and between hosts.
}
\keyword{internal}
