% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_transmit_infection.R
\name{check_and_transmit_infection}
\alias{check_and_transmit_infection}
\title{Attempt transmission within a room}
\usage{
check_and_transmit_infection(
  agents,
  room,
  current_time,
  delta_t,
  growth_rate_nasal_mean,
  growth_rate_nasal_sd,
  growth_rate_serum_mean,
  growth_rate_serum_sd,
  clearance_rate_mean,
  clearance_rate_sd,
  stochastic_noise_mean,
  stochastic_noise_sd,
  exponential_factor_mean,
  exponential_factor_sd,
  inflection_point_mean,
  inflection_point_sd,
  growth_cease_mean,
  growth_cease_sd,
  nasal_ccap_mean,
  nasal_ccap_sd,
  serum_ccap_mean,
  serum_ccap_sd,
  dose_scaling_factor,
  dose_max_efficiency,
  dose_efficiency_at_threshold,
  dose_threshold
)
}
\arguments{
\item{agents}{Data frame of all agents.}

\item{room}{Room identifier being processed.}

\item{current_time}{Current simulation time.}

\item{delta_t}{Time step in hours.}

\item{growth_rate_nasal_mean, growth_rate_nasal_sd}{Mean and SD for nasal growth rate.}

\item{growth_rate_serum_mean, growth_rate_serum_sd}{Mean and SD for serum growth rate.}

\item{clearance_rate_mean, clearance_rate_sd}{Parameters for viral clearance.}

\item{stochastic_noise_mean, stochastic_noise_sd}{Parameters for process noise.}

\item{exponential_factor_mean, exponential_factor_sd}{Parameters for waning growth.}

\item{inflection_point_mean, inflection_point_sd}{Parameters for inflection time.}

\item{growth_cease_mean, growth_cease_sd}{Parameters for growth cessation.}

\item{nasal_ccap_mean, nasal_ccap_sd}{Mean and SD of nasal carrying capacity.}

\item{serum_ccap_mean, serum_ccap_sd}{Mean and SD of serum carrying capacity.}

\item{dose_scaling_factor, dose_max_efficiency}{Parameters of dose response curve.}

\item{dose_efficiency_at_threshold, dose_threshold}{Parameters controlling dose response threshold.}
}
\value{
Updated agent data frame.
}
\description{
Checks infectious agents in a room and infects susceptible contacts based on
their nasal virus load and dose response parameters.
}
