% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_herd_incidence_dt.R
\name{collate_herd_incidence_dt}
\alias{collate_herd_incidence_dt}
\title{Fast collation of herd incidence using data.table}
\usage{
collate_herd_incidence_dt(data_path, min_incidence = 2)
}
\arguments{
\item{data_path}{Directory of herd CSV files.}

\item{min_incidence}{Trials with fewer infections than this are discarded.}
}
\value{
A data.table summarising median counts over time.
}
\description{
Equivalent to `collate_herd_incidence()` but implemented with data.table for
speed on large simulations.
}
