% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_survival_marginals.R
\name{compute_survival_marginals}
\alias{compute_survival_marginals}
\title{Compute marginal survival curves from INLA model}
\usage{
compute_survival_marginals(model, steps)
}
\arguments{
\item{model}{Fitted INLA model object.}

\item{steps}{Maximum time step to evaluate.}
}
\value{
A data frame containing survival quantiles over time.
}
\description{
Generates posterior survival probabilities for a set of time points using an
INLA fitted model.
}
