% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_calculate_phase_durations.R
\name{dt_calculate_phase_durations}
\alias{dt_calculate_phase_durations}
\title{Calculate phase durations with data.table}
\usage{
dt_calculate_phase_durations(data_path)
}
\arguments{
\item{data_path}{Directory containing result CSV files.}
}
\value{
A long `data.table` of phase durations for each animal.
}
\description{
Faster implementation of `calculate_phase_durations()` using data.table
operations.
}
