% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_calculate_phase_stats.R
\name{dt_calculate_phase_stats}
\alias{dt_calculate_phase_stats}
\title{Summarise phase durations with data.table}
\usage{
dt_calculate_phase_stats(melted_data, units = "hours")
}
\arguments{
\item{melted_data}{Long table of durations.}

\item{units}{Units to report, "hours" or "days".}
}
\value{
A data.table with summary statistics.
}
\description{
data.table implementation of `calculate_phase_stats()` for faster execution.
}
