% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_weibull.R
\name{fit_weibull}
\alias{fit_weibull}
\title{Fit a Weibull distribution from quantiles}
\usage{
fit_weibull(median, ci)
}
\arguments{
\item{median}{Median value of the distribution.}

\item{ci}{Numeric vector of length two with lower and upper quantiles.}
}
\value{
List containing `shape` and `scale`.
}
\description{
Utility for deriving shape and scale parameters given a median and
confidence interval.
}
