% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dose_table.R
\name{get_dose_table}
\alias{get_dose_table}
\title{Build dose vs incubation table}
\usage{
get_dose_table(path, periods_melted)
}
\arguments{
\item{path}{Directory of simulation results.}

\item{periods_melted}{Output from `calculate_phase_durations()`.}
}
\value{
Data frame with dose and incubation time for each animal.
}
\description{
Joins incubation periods with agent metadata to facilitate dose response
plotting.
}
