% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infect_agent.R
\name{infect_agent}
\alias{infect_agent}
\title{Infect an individual agent}
\usage{
infect_agent(
  agent_id,
  agents,
  current_time,
  initial_infection = FALSE,
  delta_t,
  dose_eff_override = NULL,
  growth_rate_nasal_mean,
  growth_rate_nasal_sd,
  growth_rate_serum_mean,
  growth_rate_serum_sd,
  clearance_rate_mean,
  clearance_rate_sd,
  stochastic_noise_mean,
  stochastic_noise_sd,
  exponential_factor_mean,
  exponential_factor_sd,
  inflection_point_mean,
  inflection_point_sd,
  growth_cease_mean,
  growth_cease_sd,
  nasal_ccap_mean,
  nasal_ccap_sd,
  serum_ccap_mean,
  serum_ccap_sd,
  infector_id = NULL
)
}
\arguments{
\item{agent_id}{Index of the agent to infect.}

\item{agents}{Data frame of all agents.}

\item{current_time}{Current simulation time.}

\item{initial_infection}{Logical indicating donor introduction.}

\item{delta_t}{Simulation time step.}

\item{dose_eff_override}{Override value for dose efficiency when infected by contact.}

\item{growth_rate_nasal_mean, growth_rate_nasal_sd}{Mean and SD of nasal growth rate.}

\item{growth_rate_serum_mean, growth_rate_serum_sd}{Mean and SD of serum growth rate.}

\item{clearance_rate_mean, clearance_rate_sd}{Parameters for clearance rate.}

\item{stochastic_noise_mean, stochastic_noise_sd}{Parameters controlling stochastic noise.}

\item{exponential_factor_mean, exponential_factor_sd}{Parameters for waning growth exponent.}

\item{inflection_point_mean, inflection_point_sd}{Parameters for growth inflection point.}

\item{growth_cease_mean, growth_cease_sd}{Mean and SD for growth cessation time.}

\item{nasal_ccap_mean, nasal_ccap_sd}{Mean and SD of nasal carrying capacity.}

\item{serum_ccap_mean, serum_ccap_sd}{Mean and SD of serum carrying capacity.}

\item{infector_id}{ID of the infecting agent if applicable.}
}
\value{
Updated agent data frame.
}
\description{
Sets initial infection parameters and virus dynamics for a given agent. When
not an initial infection the dose efficiency override is applied.
}
