% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_abm_agents.R
\name{initialize_abm_agents}
\alias{initialize_abm_agents}
\title{Initialise room-to-room agents}
\usage{
initialize_abm_agents(
  num_infected,
  num_non_infected_per_room,
  num_rooms,
  nasal_threshold_mean,
  serum_threshold_mean,
  nasal_threshold_sd,
  serum_threshold_sd,
  infect_threshold_mean,
  infect_threshold_sd
)
}
\arguments{
\item{num_infected}{Number of initially infected donors.}

\item{num_non_infected_per_room}{Number of naive contacts per room.}

\item{num_rooms}{Total rooms in the experiment.}

\item{nasal_threshold_mean, serum_threshold_mean}{Means for detection thresholds.}

\item{nasal_threshold_sd, serum_threshold_sd}{Standard deviations for thresholds.}

\item{infect_threshold_mean, infect_threshold_sd}{Parameters for infectious threshold.}
}
\value{
Data frame of initialized agents.
}
\description{
Creates donor and contact agents for a room-to-room experiment with
random thresholds and parameters.
}
