% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_animal_status.R
\name{plot_animal_status}
\alias{plot_animal_status}
\title{Plot individual infection status timelines}
\usage{
plot_animal_status(model_result)
}
\arguments{
\item{model_result}{List returned by a simulation function.}
}
\value{
A ggplot object.
}
\description{
Creates a Gantt style plot showing the infection status of each animal over
time from a single simulation result.
}
