% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparison_epicurves.R
\name{plot_comparison_epicurves}
\alias{plot_comparison_epicurves}
\title{Compare epidemic curves between scenarios}
\usage{
plot_comparison_epicurves(clinical_scenario, preclinical_scenario)
}
\arguments{
\item{clinical_scenario}{Data frame from `collate_herd_incidence_dt()`.}

\item{preclinical_scenario}{Data frame from another scenario.}
}
\value{
A ggplot object.
}
\description{
Plots median infected counts for clinical and preclinical scenarios on the
same graph.
}
