% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diverging_infections.R
\name{plot_diverging_infections}
\alias{plot_diverging_infections}
\title{Plot diverging infection counts}
\usage{
plot_diverging_infections(
  clinical_scenario,
  preclinical_scenario,
  peak_time = 168
)
}
\arguments{
\item{clinical_scenario}{Data from `collate_herd_incidence_dt()`.}

\item{preclinical_scenario}{Data from the alternative scenario.}

\item{peak_time}{Maximum time to display (hours).}
}
\value{
A ggplot object.
}
\description{
Creates a bar plot showing differences between clinical and preclinical
scenarios until a specified time point.
}
