% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dose.R
\name{plot_dose}
\alias{plot_dose}
\title{Plot dose versus incubation time}
\usage{
plot_dose(dose_plt, legend_title = "Animals")
}
\arguments{
\item{dose_plt}{Data frame created by `get_dose_table()`.}

\item{legend_title}{Title for the legend.}
}
\value{
A ggplot object.
}
\description{
Visualises the relationship between inoculation dose and resulting incubation
time using hex binning and a GAM smoother.
}
