% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_herd_curve.R
\name{plot_herd_curve}
\alias{plot_herd_curve}
\title{Plot infection and recovery counts}
\usage{
plot_herd_curve(model_result)
}
\arguments{
\item{model_result}{Result list from a simulation.}
}
\value{
A ggplot object.
}
\description{
Creates a simple line plot showing numbers infected and recovered over time
for a single simulation.
}
