% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_herd_curve_mult.R
\name{plot_herd_curve_mult}
\alias{plot_herd_curve_mult}
\title{Plot epidemic curves with uncertainty bands}
\usage{
plot_herd_curve_mult(infections_trend)
}
\arguments{
\item{infections_trend}{Data frame of summary statistics.}
}
\value{
A ggplot object.
}
\description{
Takes summary output from `collate_herd_incidence_dt()` and plots median and
quantile ribbons for infected and recovered counts.
}
