% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individual_virus_loads.R
\name{plot_individual_virus_loads}
\alias{plot_individual_virus_loads}
\title{Plot virus loads for random animals}
\usage{
plot_individual_virus_loads(model_result, sample_n = NULL)
}
\arguments{
\item{model_result}{Simulation result list.}

\item{sample_n}{Number of animals to sample.}
}
\value{
A ggplot object.
}
\description{
Selects a sample of animals from a simulation result and plots their nasal
and serum virus trajectories.
}
