% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_marginals.R
\name{plot_survival_marginals}
\alias{plot_survival_marginals}
\title{Plot survival curve marginals}
\usage{
plot_survival_marginals(
  samples,
  x_max = 14,
  xlabel = "Incubation Phase Duration",
  ylabel = "Probability"
)
}
\arguments{
\item{samples}{Data frame of survival quantiles.}

\item{x_max}{Maximum x axis in days.}

\item{xlabel}{Label for x axis.}

\item{ylabel}{Label for y axis.}
}
\value{
A ggplot object.
}
\description{
Displays median and credible intervals for survival probabilities estimated
by `compute_survival_marginals()`.
}
