% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_final_results.R
\name{process_final_results}
\alias{process_final_results}
\title{Postprocess final agent results}
\usage{
process_final_results(agents)
}
\arguments{
\item{agents}{Data frame of agent trajectories.}
}
\value{
Updated data frame.
}
\description{
Adjusts infectious start times so that animals with no clinical signs have
`NA` for `infectious_t`.
}
