% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_herd_config.R
\name{read_herd_config}
\alias{read_herd_config}
\title{Read within-herd configuration}
\usage{
read_herd_config(config_file = NULL, ...)
}
\arguments{
\item{config_file}{Path to config YAML.}

\item{...}{Named overrides.}
}
\value{
Configuration list.
}
\description{
Reads a YAML configuration file for within herd simulations. Additional
named arguments override values in the file.
}
