% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_R2R_config.R
\name{read_R2R_config}
\alias{read_R2R_config}
\title{Read room-to-room configuration}
\usage{
read_R2R_config(config_file = NULL, ...)
}
\arguments{
\item{config_file}{Path to a configuration YAML file.}

\item{...}{Named values to override in the configuration.}
}
\value{
A list of configuration parameters.
}
\description{
Loads a YAML configuration file and allows overriding parameters via
additional arguments.
}
