% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_farm_movements.R
\name{simulate_farm_movements}
\alias{simulate_farm_movements}
\title{Move animals between farms}
\usage{
simulate_farm_movements(graph, farm_status, config, time)
}
\arguments{
\item{graph}{igraph network of farms.}

\item{farm_status}{Current farm status list.}

\item{config}{Simulation parameter list.}

\item{time}{Current time step.}
}
\value{
List with updated `farm_status` and a data frame of movement events.
}
\description{
Simulates scheduled movements of animals along the edges of the farm network.
}
