% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_herd_contacts_and_transmissions.R
\name{simulate_herd_contacts_and_transmissions}
\alias{simulate_herd_contacts_and_transmissions}
\title{Simulate contacts and transmission events within a herd}
\usage{
simulate_herd_contacts_and_transmissions(
  agents,
  current_time,
  contact_rate,
  delta_t,
  growth_rate_nasal_mean,
  growth_rate_nasal_sd,
  growth_rate_serum_mean,
  growth_rate_serum_sd,
  clearance_rate_mean,
  clearance_rate_sd,
  stochastic_noise_mean,
  stochastic_noise_sd,
  exponential_factor_mean,
  exponential_factor_sd,
  inflection_point_mean,
  inflection_point_sd,
  growth_cease_mean,
  growth_cease_sd,
  nasal_ccap_mean,
  nasal_ccap_sd,
  serum_ccap_mean,
  serum_ccap_sd,
  dose_scaling_factor,
  dose_max_efficiency,
  dose_efficiency_at_threshold,
  dose_threshold,
  preclin_infect
)
}
\arguments{
\item{agents}{Data frame of agents.}

\item{current_time}{Current time step.}

\item{contact_rate}{Mean number of contacts per animal per time step.}

\item{delta_t}{Simulation time step in hours.}

\item{growth_rate_nasal_mean, growth_rate_nasal_sd}{Growth rate parameters for nasal virus.}

\item{growth_rate_serum_mean, growth_rate_serum_sd}{Growth rate parameters for serum virus.}

\item{clearance_rate_mean, clearance_rate_sd}{Clearance parameters.}

\item{stochastic_noise_mean, stochastic_noise_sd}{Noise parameters.}

\item{exponential_factor_mean, exponential_factor_sd}{Exponential decay parameters.}

\item{inflection_point_mean, inflection_point_sd}{Inflection point parameters.}

\item{growth_cease_mean, growth_cease_sd}{Growth cessation parameters.}

\item{nasal_ccap_mean, nasal_ccap_sd}{Nasal carrying capacity parameters.}

\item{serum_ccap_mean, serum_ccap_sd}{Serum carrying capacity parameters.}

\item{dose_scaling_factor, dose_max_efficiency}{Dose response curve parameters.}

\item{dose_efficiency_at_threshold, dose_threshold}{Dose response threshold parameters.}

\item{preclin_infect}{Logical: infectious before clinical signs?}
}
\value{
Updated agent data frame.
}
\description{
For each infectious animal, contacts are generated and susceptible animals may
become infected according to a dose response.
}
