% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_within_herd.R
\name{simulate_within_herd}
\alias{simulate_within_herd}
\title{Simulate within-herd transmission}
\usage{
simulate_within_herd(config_file = NULL, ...)
}
\arguments{
\item{config_file}{Path to configuration file.}

\item{...}{Override values for the configuration list.}
}
\value{
List of agent metadata and final results over time.
}
\description{
Runs the stochastic within herd model for a single herd based on a YAML
configuration.
}
