% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_agent_virus_dynamics.R
\name{update_agent_virus_dynamics}
\alias{update_agent_virus_dynamics}
\title{Update virus dynamics for an agent}
\usage{
update_agent_virus_dynamics(agent, current_time, delta_t)
}
\arguments{
\item{agent}{Data frame row for an individual animal.}

\item{current_time}{Current simulation time.}

\item{delta_t}{Time step in hours.}
}
\value{
Updated agent row.
}
\description{
Applies logistic growth with clearance and noise to update viral loads for an
infected agent.
}
