% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_within_herd_step.R
\name{update_within_herd_step}
\alias{update_within_herd_step}
\title{Simulate one within herd time step}
\usage{
update_within_herd_step(agents, state_df, config, time)
}
\arguments{
\item{agents}{Data frame of herd agents.}

\item{state_df}{State history data frame.}

\item{config}{List of parameters for transmission and recovery.}

\item{time}{Current simulation time.}
}
\value{
List containing updated `agents` and `state_df`.
}
\description{
Updates infection states for all animals in a herd and records the results in
the state data frame.
}
